+++ packages/CHANGELOG.md
@@ -1,1 +1,7 @@
-# copy vuepress-theme-default (2021-04-07)
+### 2021-04-10
+
+- 删除原有 footer，新增 Footer 组件，支持备案信息展示
+
+### 2021-04-07
+
+- copy vuepress-theme-default